agente/requirements.txt
yt-dlp>=2024.0.0
openai>=1.0.0
python-dotenv>=1.0.0
